+++ demo/src/app/core/menu.ts
@@ -58,7 +58,9 @@ export const allMenus = [
                     { name: 'Dropdown', zhName: '下拉菜单', routePath: 'dropdown' },
                     { name: 'Cascader', zhName: '级联选择菜单', routePath: 'cascader' },
                     { name: 'Property Operation', zhName: '属性操作', routePath: 'property-operation' },
-                    { name: 'Arrow-switcher', zhName: '上下条切换', routePath: 'arrow-switcher' }
+                    { name: 'Arrow-switcher', zhName: '上下条切换', routePath: 'arrow-switcher' },
+                    { name: 'Vote', zhName: '投票', routePath: 'vote' },
+                    { name: 'Copy', zhName: '复制', routePath: 'copy' }
                 ]
             },
             {
@@ -114,7 +116,8 @@ export const allMenus = [
                     { name: 'Tree', zhName: '树', routePath: 'tree' },
                     { name: 'Stepper', zhName: '步骤条', routePath: 'stepper' },
                     { name: 'Progress', zhName: '进度条', routePath: 'progress' },
-                    { name: 'Markdown', zhName: 'markdown 解析', routePath: 'markdown' }
+                    { name: 'Markdown', zhName: 'markdown 解析', routePath: 'markdown' },
+                    { name: 'Result', zhName: '结果页', routePath: 'result' }
                 ]
             },
             {