+++ src/module.ts
@@ -53,6 +53,8 @@ import { ThyFlexibleTextModule } from './flexible-text';
 import { ThyDragDropModule } from './drag-drop/module';
 import { ThySelectCommonModule } from './core/select/module';
 import { ThySkeletonModule } from './skeleton';
+import { ThyVoteModule } from './vote';
+import { ThyResultModule } from './result';
 
 const IMPORT_EXPORT_MODULES = [
     BrowserAnimationsModule,
@@ -109,7 +111,9 @@ const IMPORT_EXPORT_MODULES = [
     ThyFlexibleTextModule,
     ThyDragDropModule,
     ThySelectCommonModule,
-    ThySkeletonModule
+    ThySkeletonModule,
+    ThyVoteModule,
+    ThyResultModule
 ];
 
 @NgModule({