+++ src/stepper/stepper-button.directive.ts
@@ -1,4 +1,3 @@
-
 import { Directive, Input, HostListener } from '@angular/core';
 import { ThyStepperComponent } from './stepper.component';
 
@@ -7,29 +6,23 @@ import { ThyStepperComponent } from './stepper.component';
     selector: '[thyStepperNext]'
 })
 export class ThyStepperNextDirective {
-
-    constructor(private stepper: ThyStepperComponent) {
-    }
+    constructor(private stepper: ThyStepperComponent) {}
 
     @HostListener('click', ['$event'])
     click($event: any) {
         this.stepper.next();
     }
-
 }
 
 /** Button that moves to the previous step in a stepper workflow. */
 @Directive({
-    selector: '[thyStepperPrevious]',
+    selector: '[thyStepperPrevious]'
 })
 export class ThyStepperPreviousDirective {
-
-    constructor(private stepper: ThyStepperComponent) {
-    }
+    constructor(private stepper: ThyStepperComponent) {}
 
     @HostListener('click', ['$event'])
     click($event: any) {
         this.stepper.previous();
     }
-
 }