+++ src/vote/index.ts
@@ -1,0 +1,2 @@
+export * from './vote.module';
+export * from './vote.component';